+++ apps/admin/src/components/Session/Empty.tsx
@@ -1,7 +1,6 @@
 import { FC } from 'react'
 import { Box, Typography } from '@material-ui/core'
 import { CreateButton, useListContext } from 'react-admin'
-import { CollaboratorChallengeDialogWithButton } from '.'
 
 const Empty = () => {
   const { basePath } = useListContext()
@@ -15,7 +14,6 @@ const Empty = () => {
         さっそく、新しいシナリオを登録してみましょう
       </Typography>
       <CreateButton basePath={basePath} size="large" variant="contained" />
-      <CollaboratorChallengeDialogWithButton />
     </Box>
   )
 }
