+++ apps/admin/src/components/Session/index.tsx
@@ -1,4 +1,4 @@
-import { FC, useCallback, useState } from 'react'
+import { FC } from 'react'
 import {
   Datagrid,
   List,
@@ -16,38 +16,16 @@ import {
   DeleteButton,
   ToolbarProps,
   Show,
-  TabbedShowLayout,
-  Tab,
-  ShowProps,
-  useShowContext,
-  TextInput,
-  required,
-  email,
-  TopToolbar,
-  CreateButton,
-  ListActionsProps,
-  sanitizeListRestProps,
-  Button as RaButton
+  ShowProps
 } from 'react-admin'
 import EditForm from './Edit'
 import CreateForm from './Create'
 import EmbeddedScriptDialog from './EmbeddedScriptDialog'
 import { Session } from '@botui/types'
 import PreviewDialog from './PreviewDialog'
-import { useFormState, Form } from 'react-final-form'
+import { useFormState } from 'react-final-form'
 import Empty from './Empty'
-import {
-  Chip,
-  makeStyles,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogContentText,
-  DialogActions,
-  Button
-} from '@material-ui/core'
-import PersonAdd from '@material-ui/icons/PersonAdd'
-import { Auth } from 'aws-amplify'
+import { ShowInner } from './Show'
 
 const EditToolbar: FC<Omit<ToolbarProps, 'width'>> = (props) => {
   const { values: session } = useFormState<Session>()
@@ -60,95 +38,11 @@ const EditToolbar: FC<Omit<ToolbarProps, 'width'>> = (props) => {
   )
 }
 
-const ListActions: FC<ListActionsProps> = (props) => {
-  return (
-    <TopToolbar {...sanitizeListRestProps(props)}>
-      <CollaboratorChallengeDialogWithButton />
-      <CreateButton label="シナリオを作る" />
-    </TopToolbar>
-  )
-}
-
-export const CollaboratorChallengeDialogWithButton: FC = () => {
-  const [open, setOpen] = useState(false)
-  const refresh = useRefresh()
-  const handleClickOpen = () => {
-    setOpen(true)
-  }
-  const handleClose = () => {
-    setOpen(false)
-  }
-  const challenge = useCallback(
-    ({ code }: { code: string }) => {
-      Auth.currentUserInfo().then((user) => {
-        console.log(user)
-        fetch('http://localhost:3333/api/collaborator/challenge', {
-          method: 'POST',
-          headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify({
-            email: user.attributes.email,
-            code,
-            userId: user.username
-          })
-        }).then(() => {
-          handleClose()
-          refresh()
-        })
-      })
-    },
-    [refresh]
-  )
-
-  return (
-    <>
-      <RaButton
-        onClick={handleClickOpen}
-        startIcon={<PersonAdd />}
-        label="招待コードを入力"
-      />
-      <Dialog onClose={handleClose} open={open}>
-        <Form
-          onSubmit={challenge}
-          render={({ invalid, submitting, handleSubmit }) => (
-            <>
-              <DialogTitle>招待コードを入力</DialogTitle>
-              <DialogContent>
-                <DialogContentText>
-                  招待コードを受け取っている場合には、コードを入力することで共同編集が可能です。
-                </DialogContentText>
-                <TextInput
-                  source="code"
-                  label="招待コード"
-                  validate={[required()]}
-                  fullWidth
-                />
-              </DialogContent>
-              <DialogActions>
-                <Button
-                  disabled={invalid || submitting}
-                  onClick={handleSubmit}
-                  color="primary"
-                >
-                  確定
-                </Button>
-              </DialogActions>
-            </>
-          )}
-        ></Form>
-      </Dialog>
-    </>
-  )
-}
-
 export const SessionList: FC = (props) => {
   return (
     <List
       {...props}
       bulkActionButtons={false}
-      actions={<ListActions />}
       empty={<Empty />}
     >
       <Datagrid>
@@ -191,140 +85,10 @@ export const SessionEdit: FC = (props) => {
   )
 }
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexWrap: 'wrap',
-    '& > *': {
-      margin: theme.spacing(0.5)
-    }
-  }
-}))
-
 export const SessionShow: FC<ShowProps> = (props) => {
-  const classes = useStyles()
-  const refresh = useRefresh()
-  const remove = useCallback(
-    (sessionId: string, email: string) => {
-      if (!window.confirm('共同編集者から外しますか？')) return
-      fetch('http://localhost:3333/api/collaborator/remove', {
-        method: 'POST',
-        headers: {
-          Accept: 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-          sessionId,
-          email
-        })
-      }).then(() => refresh())
-    },
-    [refresh]
-  )
-
   return (
     <Show {...props}>
-      <TabbedShowLayout>
-        <Tab label="共同編集者">
-          <FunctionField<Session>
-            source="collaborators"
-            render={(record) => (
-              <div className={classes.root}>
-                {record?.collaborators?.map((email) => (
-                  <Chip
-                    key={email}
-                    label={email}
-                    onDelete={() => remove(record.id, email)}
-                    color="primary"
-                  />
-                ))}
-                {record?.collaboratorInvitations?.items
-                  ?.filter(({ status }) => status === 'inviting')
-                  .map(({ email, id, sessionId }) => (
-                    <Chip
-                      key={id}
-                      label={`招待中: ${email}`}
-                      onDelete={() => remove(sessionId, email)}
-                      variant="outlined"
-                      color="primary"
-                    />
-                  ))}
-                <InviteDialogWithChipButton />
-              </div>
-            )}
-          />
-        </Tab>
-      </TabbedShowLayout>
+      <ShowInner />
     </Show>
-  )
-}
-
-const InviteDialogWithChipButton: FC<ShowProps> = (props) => {
-  const [open, setOpen] = useState(false)
-  const session = useShowContext<Session>(props)
-  const refresh = useRefresh()
-  const handleClickOpen = () => {
-    setOpen(true)
-  }
-  const handleClose = () => {
-    setOpen(false)
-  }
-  const invite = useCallback(
-    ({ email }: { email: string }) => {
-      fetch('http://localhost:3333/api/collaborator/invite', {
-        method: 'POST',
-        headers: {
-          Accept: 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-          sessionTitle: session.record?.title,
-          sessionId: session.record?.id,
-          email
-        })
-      }).then(() => {
-        handleClose()
-        refresh()
-      })
-    },
-    [refresh, session.record?.id, session.record?.title]
-  )
-
-  return (
-    <>
-      <Chip label="招待する" clickable onClick={handleClickOpen} />
-      <Dialog onClose={handleClose} open={open}>
-        <Form
-          onSubmit={invite}
-          render={({ invalid, submitting, handleSubmit }) => (
-            <>
-              <DialogTitle>共同編集者を招待</DialogTitle>
-              <DialogContent>
-                <DialogContentText>
-                  入力したメールアドレスに招待リンク付きのメールが送信されます。招待リンクをクリックすると共同編集者に登録されます。
-                </DialogContentText>
-                <TextInput
-                  source="email"
-                  type="email"
-                  label="通知用メールアドレス"
-                  validate={[required(), email()]}
-                  fullWidth
-                />
-              </DialogContent>
-              <DialogActions>
-                <Button
-                  disabled={invalid || submitting}
-                  onClick={handleSubmit}
-                  color="primary"
-                >
-                  招待する
-                </Button>
-              </DialogActions>
-            </>
-          )}
-        ></Form>
-      </Dialog>
-    </>
   )
 }
